pas_handler_v3: change shouldSendStatus1sTime type from uint16_t to uint32_t to prevent overflowing

+++ firmware/stm32/helpers/pas_handler_v3/Core/Src/uart.c
@@ -22,7 +22,7 @@ char *currentMessage = "eskl_initCompl\r\n";
 
 bool shouldSendStatus = false;
 
-uint16_t shouldSendStatus1sTime = 0;
+uint32_t shouldSendStatus1sTime = 0;
 bool shouldSendStatus1s = false;
 
 void handleRxInterrupt(UART_HandleTypeDef *huart, uint16_t size)
